requirements.txt
awscli @ ./dependencies/awscli-1.29.21-py3-none-any.whl
blinker==1.6.2
boto3 @ ./dependencies/boto3-1.28.21-py3-none-any.whl
botocore @ ./dependencies/botocore-1.31.21-py3-none-any.whl
click==8.1.7
colorama==0.4.4
docutils==0.16
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
gunicorn==21.2.0
